Add general domain prior
This change adds a data structure encapsulating general prior data in
the form of eigenvalues and eigenvectors.

The current default behaviour is to assume a unforimly gridded rectangle
with a periodic boundary. This case should also be encapsulated by the
new data structure, with the behaviour being enabled by a flag argument
to the chain initialiser.